refactor: address remaining review comments
- Remove redundant 'as u64' cast in default calculation
- Enhance error message to include limit name (max_module_size)
- Expand safety documentation for from_precompiled explaining security trade-off

+++ crates/runtime/src/logic.rs
@@ -170,7 +170,7 @@ impl Default for VMLimits {
         }
 
         Self {
-            max_module_size: DEFAULT_MAX_MODULE_SIZE_MIB as u64 * u64::from(ONE_MIB),
+            max_module_size: DEFAULT_MAX_MODULE_SIZE_MIB * u64::from(ONE_MIB),
             max_memory_pages: ONE_KIB,
             max_stack_size: DEFAULT_MAX_STACK_SIZE_KIB * ONE_KIB as usize,
             max_registers: DEFAULT_MAX_REGISTERS,
